+++ packages/terra-notification-dialog/tests/wdio/notification-dialog-spec.js
@@ -174,3 +174,37 @@ Terra.viewports('tiny', 'medium', 'large').forEach((viewport) => {
     });
   });
 });
+
+describe('NotificationDialog with additional focus trap sources within a modal manager', () => {
+  before(() => {
+    browser.url('/#/raw/tests/terra-notification-dialog/notification-dialog/notification-dialog-on-modal-manager');
+    browser.setViewportSize(Terra.viewports('medium')[0]);
+    browser.click('#disclose-modal');
+  });
+
+  it('opens popup', () => {
+    browser.click('#popup-in-modal');
+  });
+
+  Terra.should.matchScreenshot('Open popup', { selector: '[class*="abstract-modal"]' });
+
+  it('dismisses popup', () => {
+    browser
+      .moveToObject('#root', 300, 300)
+      .leftClick();
+  });
+
+  Terra.should.matchScreenshot('Close popup', { selector: '[class*="abstract-modal"]' });
+
+  it('opens notification dialog', () => {
+    browser.click('#trigger-notification-dialog');
+  });
+
+  Terra.should.matchScreenshot('Open notification dialog', { selector: '[class*="abstract-modal"]' });
+
+  it('dismiss notification dialog', () => {
+    browser.click('[class*="notification-dialog-inner-wrapper"] button:last-child');
+  });
+
+  Terra.should.matchScreenshot('Close notification dialog', { selector: '[class*="abstract-modal"]' });
+});